Update user-handler.php
Check if only pre-determined users are already set in the arguments to the query, in which case some upstream filter has already set which specific users to show and we don't need to build a query for the authors. 

Co-Authors Plus does this by setting only the include argument with user_id 0 in the filter quick_edit_dropdown_authors_args to suppress the regular author dropdown and show only the co-authors selection box.